# go-backend/docs/swagger.yaml
basePath: /
definitions:
...
  services.ActivateCarRequest:
    properties:
      user_id:
        type: integer
    type: object
  services.AddCarHistoryRequest:
    properties:
      action_type:
        type: string
      comment:
        type: string
      field_name:
        type: string
      metadata:
        type: object
      new_value:
        type: string
      old_value:
        type: string
      user_id:
        type: integer
    type: object
...
  services.AllCarsHistoryItem:
    properties:
      action_type:
        type: string
      car_brand:
        type: string
      car_id:
        type: integer
      car_number:
        type: string
      comment:
        type: string
      company:
        type: string
      created_at:
        type: string
      id:
        type: integer
      organization:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
  services.CarCurrentStatus:
    properties:
      car_id:
        type: integer
      entry_time:
        type: string
      last_exit_time:
        type: string
      territory_status:
        type: integer
    type: object
  services.CarHistoryItemResponse:
    properties:
      action_type:
        type: string
      car_brand:
        type: string
      car_id:
        type: integer
      car_number:
        type: string
      comment:
        type: string
      company:
        type: string
      created_at:
        type: string
      field_name:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      metadata:
        type: object
      middle_name:
        type: string
      new_value:
        type: string
      old_value:
        type: string
      organization:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
  services.CarUnloadPlaceInfo:
    properties:
      car_id:
        type: integer
      unload_place_id:
        type: integer
      unload_place_name:
        type: string
    type: object
...
  services.CheckActiveCarResponse:
    properties:
      active:
        type: boolean
      application_id:
        type: integer
      application_number:
        type: string
      car_brand:
        type: string
      car_id:
        type: integer
      car_number:
        type: string
      company_name:
        type: string
      entry_date_to:
        type: string
      entry_time_to:
        type: string
      organization_name:
        type: string
    type: object
...
  services.DeactivateCarRequest:
    properties:
      status:
        type: integer
      user_id:
        type: integer
    type: object
...
  services.RestoreCarRequest:
    properties:
      user_id:
        type: integer
    type: object
...
  services.TableCarResponse:
    properties:
      application_id:
        type: integer
      car_brand:
        type: string
      car_number:
        type: string
      company:
        type: string
      company_id:
        type: integer
      entry_date_to:
        type: string
      entry_time_from:
        type: string
      entry_time_to:
        type: string
      id:
        type: integer
      organization:
        type: string
      organization_id:
        type: integer
      status:
        type: integer
      territory_entry_time:
        type: string
      territory_status:
        type: integer
      unload_place:
        type: string
      unload_places:
        items:
          type: string
        type: array
    type: object
...
  services.UpdateTerritoryStatusRequest:
    properties:
      territory_status:
        type: integer
      user_id:
        type: integer
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API системы управления пропусками (Бюро пропусков)
  title: Systemburo API
  version: "1.0"
paths:
...
  /cars/{id}/activate:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      - description: Данные активации
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.ActivateCarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Активация автомобиля (ввод в работу)
      tags:
      - cars
  /cars/{id}/deactivate:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      - description: Данные деактивации
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.DeactivateCarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Деактивация автомобиля
      tags:
      - cars
  /cars/{id}/history:
    get:
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CarHistoryItemResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Получение истории конкретного автомобиля
      tags:
      - cars
    post:
      consumes:
      - application/json
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      - description: Данные записи истории
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.AddCarHistoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Добавление записи в историю автомобиля
      tags:
      - cars
  /cars/{id}/restore:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      - description: Данные восстановления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.RestoreCarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Восстановление удалённого автомобиля
      tags:
      - cars
  /cars/{id}/territory-status:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      - description: Новый территориальный статус
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.UpdateTerritoryStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновление статуса нахождения автомобиля на территории
      tags:
      - cars
  /cars/active-for-tables:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.TableCarResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Получение активных машин для таблиц
      tags:
      - cars
  /cars/check-active:
    get:
      parameters:
      - description: Номер автомобиля
        in: query
        name: car_number
        required: true
        type: string
      - description: Марка автомобиля
        in: query
        name: car_brand
        required: true
        type: string
      - description: ID организации
        in: query
        name: organization_id
        type: integer
      - description: ID компании
        in: query
        name: company_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.CheckActiveCarResponse'
      security:
      - BearerAuth: []
      summary: Проверка наличия активной машины по параметрам
      tags:
      - cars
  /cars/fact-for-tables:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.TableCarResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Получение машин «по факту» для таблиц
      tags:
      - cars
  /cars/fact-unload-places:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CarUnloadPlaceInfo'
            type: array
      security:
      - BearerAuth: []
      summary: Получение связей машин «по факту» с местами разгрузки
      tags:
      - cars
  /cars/history/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.AllCarsHistoryItem'
            type: array
      security:
      - BearerAuth: []
      summary: Получение истории въездов/выездов всех автомобилей
      tags:
      - cars
  /cars/history/current-status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CarCurrentStatus'
            type: array
      security:
      - BearerAuth: []
      summary: Получение текущего территориального статуса активных автомобилей
      tags:
      - cars
  /cars/history/unified:
    get:
      parameters:
      - description: Номер автомобиля
        in: query
        name: car_number
        required: true
        type: string
      - description: Марка автомобиля
        in: query
        name: car_brand
        required: true
        type: string
      - description: ID организации
        in: query
        name: organization_id
        type: integer
      - description: ID компании
        in: query
        name: company_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CarHistoryItemResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Получение объединённой истории для всех машин с одинаковыми параметрами
      tags:
      - cars
  /cars/unload-places:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CarUnloadPlaceInfo'
            type: array
      security:
      - BearerAuth: []
      summary: Получение связей активных машин с местами разгрузки
      tags:
      - cars
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"